removing not needed stuff from country details page
