--- constants/countries-detail.js
@@ -1,18 +1,0 @@
-const TABS_COUNTRIES_DETAIL = [
-  {
-    label: 'overview',
-    value: 'overview'
-  },
-  {
-    label: 'vpas-documentation',
-    value: 'documentation'
-  },
-  {
-    label: 'observations',
-    value: 'observations'
-  }
-];
-
-export {
-  TABS_COUNTRIES_DETAIL
-};
