+++ modules/countries-detail.js
@@ -7,10 +7,6 @@ const GET_COUNTRY_SUCCESS = 'GET_COUNTRY_SUCCESS';
 const GET_COUNTRY_ERROR = 'GET_COUNTRY_ERROR';
 const GET_COUNTRY_LOADING = 'GET_COUNTRY_LOADING';
 
-const GET_COUNTRY_OBSERVATIONS_SUCCESS = 'GET_COUNTRY_OBSERVATIONS_SUCCESS';
-const GET_COUNTRY_OBSERVATIONS_ERROR = 'GET_COUNTRY_OBSERVATIONS_ERROR';
-const GET_COUNTRY_OBSERVATIONS_LOADING = 'GET_COUNTRY_OBSERVATIONS_LOADING';
-
 const GET_COUNTRY_LINKS_SUCCESS = 'GET_COUNTRY_LINKS_SUCCESS';
 const GET_COUNTRY_LINKS_ERROR = 'GET_COUNTRY_LINKS_ERROR';
 const GET_COUNTRY_LINKS_LOADING = 'GET_COUNTRY_LINKS_LOADING';
@@ -25,11 +21,6 @@ const initialState = {
   loading: false,
   error: false,
   documentation: {
-    data: {},
-    loading: false,
-    error: false
-  },
-  observations: {
     data: {},
     loading: false,
     error: false
@@ -60,22 +51,6 @@ export default function reducer(state = initialState, action) {
     case GET_COUNTRY_LOADING: {
       return Object.assign({}, state, { loading: true, error: false });
     }
-
-    case GET_COUNTRY_OBSERVATIONS_SUCCESS: {
-      const observations = Object.assign({}, state.observations, {
-        data: action.payload, loading: false, error: false
-      });
-      return Object.assign({}, state, { observations });
-    }
-    case GET_COUNTRY_OBSERVATIONS_ERROR: {
-      const observations = Object.assign({}, state.observations, { error: true, loading: false });
-      return Object.assign({}, state, { observations });
-    }
-    case GET_COUNTRY_OBSERVATIONS_LOADING: {
-      const observations = Object.assign({}, state.observations, { loading: true, error: false });
-      return Object.assign({}, state, { observations });
-    }
-
     case GET_COUNTRY_LINKS_SUCCESS: {
       const links = Object.assign({}, state.links, {
         data: action.payload, loading: false, error: false
@@ -160,63 +135,6 @@ export function getCountry(id) {
         // Fetch from server ko -> Dispatch error
         dispatch({
           type: GET_COUNTRY_ERROR,
-          payload: err.message
-        });
-      });
-  };
-}
-
-
-/* Action creators */
-export function getCountryObservations(id) {
-  return (dispatch, getState) => {
-    const { user, language } = getState();
-
-    // Waiting for fetch from server -> Dispatch loading
-    dispatch({ type: GET_COUNTRY_OBSERVATIONS_LOADING });
-
-    const includeFields = [
-      'severity',
-      'subcategory',
-      'subcategory.category',
-      'observation-report',
-      'observation-documents'
-    ];
-
-    const lang = language === 'zh' ? 'zh-CN' : language;
-
-    const queryParams = queryString.stringify({
-      ...!!includeFields.length && { include: includeFields.join(',') },
-      locale: lang,
-      'filter[country_id]': id
-    });
-
-
-    return fetch(`${process.env.OTP_API}/observations/?${queryParams}`, {
-      method: 'GET',
-      headers: {
-        'Content-Type': 'application/json',
-        'OTP-API-KEY': process.env.OTP_API_KEY,
-        Authorization: user.token ? `Bearer ${user.token}` : undefined
-      }
-    })
-      .then((response) => {
-        if (response.ok) return response.json();
-        throw new Error(response.statusText);
-      })
-      .then((country) => {
-        // Fetch from server ok -> Dispatch country and deserialize the data
-        const dataParsed = JSONA.deserialize(country);
-
-        dispatch({
-          type: GET_COUNTRY_OBSERVATIONS_SUCCESS,
-          payload: dataParsed
-        });
-      })
-      .catch((err) => {
-        // Fetch from server ko -> Dispatch error
-        dispatch({
-          type: GET_COUNTRY_OBSERVATIONS_ERROR,
           payload: err.message
         });
       });
